requirements.txt
beautifulsoup4>=4.12.3
bs4>=0.0.2
feedparser>=6.0.11
nltk>=3.8.1
nostr>=0.0.2
pip>=23.3.1
pipenv>=2023.10.24
python-dotenv>=1.0.1
scikit-learn>=1.4.1.post1